requirements.txt
xlrd
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
cycler==0.10.0
Django==2.1
django-cors-headers==3.0.2
django-multiselectfield==0.1.8
django-redis==4.10.0
djangorestframework==3.10.1
DjangoUeditor==1.8.143
dwebsocket==0.5.10
idna==2.8
kiwisolver==1.1.0
numpy==1.17.0
Pillow==6.1.0
pycryptodome==3.8.2
pycryptodomex==3.7.2
PyMySQL==0.9.3
pyparsing==2.4.2
python-alipay-sdk==1.10.1
python-dateutil==2.8.0
pytz==2019.1
redis==3.2.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
urllib3==1.25.3
xmltodict==0.12.0
yunpian-python-sdk==1.0.0
